+++ ui/src/app/core/components/information/information.component.ts
@@ -26,7 +26,6 @@ export class InformationComponent {
   public title = this.modalData.title
   public subtitle = this.modalData.subtitle
   public message = this.modalData.message
-  public message2 = this.modalData.markdownMessage2
   public ctaButtonLabel = this.modalData.ctaButtonLabel
   public ctaButtonLink = this.modalData.ctaButtonLink
   public faIconClass = this.modalData.faIconClass