+++ ui/src/app/modules/platform-tools/docker/container-restart/container-restart.component.ts
@@ -51,12 +51,6 @@ export class ContainerRestartComponent implements OnInit, OnDestroy {
       void this.$settings.getAppSettings().catch(() => { /* do nothing */ })
     })
 
-    // Check if already connected and initialize immediately
-    if (this.io.socket.connected) {
-      this.io.socket.emit('monitor-server-status')
-      void this.$settings.getAppSettings().catch(() => { /* do nothing */ })
-    }
-
     try {
       await this.$api.put('/platform-tools/docker/restart-container', {})
       this.checkIfServerUp()