+++ ui/src/app/modules/status/widgets/uptime-widget/uptime-widget.component.ts
@@ -38,11 +38,6 @@ export class UptimeWidgetComponent implements OnInit {
         this.getServerUptimeInfo()
       }
     })
-
-    // Fetch initial data if already connected
-    if (this.io.socket.connected) {
-      this.getServerUptimeInfo()
-    }
   }
 
   private getServerUptimeInfo(): void {